feat(table): add support for aligned subtable columns (#4613)
## What does this do?
Adds additional props to the Table API to allow for automatic column
alignment of the subtable columns with the parent table. This is done by
setting the subtable's column width to the same value as the parent
column width.

## Screenshot / Video


https://github.com/user-attachments/assets/1fdf992b-5eba-4940-8783-41858add1847



## Relevant tickets / Documentation
- Ticket: [RET-830: Alignment of table <-> subtable
columns](https://linear.app/marshmallow/issue/RET-830/alignment-of-table-subtable-columns)
- closes RET-830

### How do we enable this feature?
- set `shouldAlignSubTableColumns` to `true` on your parent table. This
will set the widths of your subtable columns to match the widths of the
parent table.